Move gists/gist5707298 to parserutil.
